pyproject.toml
[project]
name = "django-ninja-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-ninja>=1.4.3",
]
